feat: 사이드바 채팅방 목록 실시간 반영 추가 #184
