+++ src/main/java/com/example/grapefield/chat/kafka/ChatParticipantEventKafkaConsumer.java
@@ -2,25 +2,20 @@ package com.example.grapefield.chat.kafka;
 
 import com.example.grapefield.chat.model.response.ChatParticipantEventResp;
 import com.example.grapefield.chat.model.response.UserChatListEventResp;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.core.type.TypeReference; // 이 import가 누락되었을 수 있음
+import com.example.grapefield.chat.service.ChatParticipantService;
+import com.example.grapefield.chat.service.UserChatListService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Component;
-
-import java.util.Map;
 
 @Component
 @Slf4j
 @RequiredArgsConstructor
 public class ChatParticipantEventKafkaConsumer {
 
-    private final RedisTemplate<String, String> redisTemplate;
-    private final SimpMessagingTemplate messagingTemplate;
-    // ObjectMapper 제거 - 더 이상 필요 없음!
+    private final ChatParticipantService participantService;
+    private final UserChatListService userChatListService;
 
     // 참여자 수 변경 이벤트 처리
     @KafkaListener(
@@ -29,16 +24,8 @@ public class ChatParticipantEventKafkaConsumer {
             containerFactory = "participantEventKafkaListenerContainerFactory"
     )
     public void handleParticipantEvent(ChatParticipantEventResp event) {
-        try {
-            if ("PARTICIPANT_CHANGE".equals(event.getType())) {
-                int count = getParticipantCount(event.getRoomIdx());
-                String topic = "/topic/chat/room/" + event.getRoomIdx() + "/participants";
-                messagingTemplate.convertAndSend(topic, count);
-                log.info("📢 [카프카] {} 브로드캐스트: {} → {}", event.getAction(), topic, count);
-            }
-        } catch (Exception e) {
-            log.error("❌ 참여자 수 카프카 이벤트 처리 실패: ", e);
-        }
+        log.info("📨 참여자 이벤트 수신: {}", event);
+        participantService.handleParticipantChange(event);
     }
 
     // 사용자 채팅방 리스트 이벤트 처리
@@ -48,31 +35,7 @@ public class ChatParticipantEventKafkaConsumer {
             containerFactory = "userEventKafkaListenerContainerFactory"
     )
     public void handleUserChatListEvent(UserChatListEventResp event) {
-        try {
-            if ("USER_CHATLIST_UPDATE".equals(event.getType())) {
-                String userTopic = "/topic/user/" + event.getUserIdx() + "/chatlist";
-
-                Map<String, Object> notifyMessage = Map.of(
-                        "action", event.getAction(),
-                        "roomIdx", event.getRoomIdx(),
-                        "timestamp", System.currentTimeMillis()
-                );
-
-                messagingTemplate.convertAndSend(userTopic, notifyMessage);
-                log.info("📢 [사용자 채팅방 리스트] 웹소켓 전송: {} → {}", userTopic, notifyMessage);
-            }
-        } catch (Exception e) {
-            log.error("❌ 사용자 채팅방 리스트 이벤트 처리 실패: ", e);
-        }
-    }
-
-
-    private int getParticipantCount(Long roomIdx) {
-        String key = "chat:room:" + roomIdx + ":participants";
-        String cached = redisTemplate.opsForValue().get(key);
-        if (cached != null) {
-            return Integer.parseInt(cached);
-        }
-        return 0;
+        log.info("📨 사용자 채팅방 리스트 이벤트 수신: {}", event);
+        userChatListService.handleChatListUpdate(event);
     }
 }
